+++ tests/testthat/test-provider-cortex.R
@@ -124,7 +124,7 @@ test_that("Cortex API requests are generated correctly", {
   req <- chat_request(p, FALSE, list(turn))
   expect_snapshot(
     req,
-    transform = function(x) gsub(ellmer_user_agent(), "<ellmer_user_agent>", x, fixed = TRUE)
+    transform = function(x) gsub(snowflake_user_agent(), "<ellmer_user_agent>", x, fixed = TRUE)
   )
   expect_snapshot(req$body$data)
 })